Changes to fix minor issue with getToken request on DatalayerServiceImpl & move DatalayerServiceImplTest.java to correct location
